3.4-django-testing/django_testing/requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
pytest-cov
model-bakery